+++ main/Main.hs
@@ -14,10 +14,11 @@ module Main where
 import Options.Applicative
 
 import              Qux.Commands
-import qualified    Qux.Commands.Build      as Build
-import qualified    Qux.Commands.Check      as Check
-import qualified    Qux.Commands.Compile    as Compile
-import qualified    Qux.Commands.Print      as Print
+import qualified    Qux.Commands.Build          as Build
+import qualified    Qux.Commands.Check          as Check
+import qualified    Qux.Commands.Compile        as Compile
+import qualified    Qux.Commands.Dependencies   as Dependencies
+import qualified    Qux.Commands.Print          as Print
 
 
 -- | Main.
@@ -26,8 +27,9 @@ main = customExecParser quxPrefs quxInfo >>= handle
 
 handle :: Options -> IO ()
 handle options = case argCommand options of
-    Build   options -> Build.handle     options
-    Check   options -> Check.handle     options
-    Compile options -> Compile.handle   options
-    Print   options -> Print.handle     options
+    Build           options -> Build.handle         options
+    Check           options -> Check.handle         options
+    Compile         options -> Compile.handle       options
+    Dependencies    options -> Dependencies.handle  options
+    Print           options -> Print.handle         options