+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 #### Upcoming in master
 
+* Added in support for import statements.
+* Added in `dependencies` command that prints out a tree of the file imports.
+
 #### v0.2.0.0
 
 * Updated usage text for `check`.